documind/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.30.6

# Data validation
pydantic==2.7.4
pydantic-settings==2.3.4

# Logging
loguru==0.7.2

# File handling
python-multipart==0.0.9
aiofiles==24.1.0

# HTTP client
httpx==0.27.2

# PDF extraction (Chapter 5)
pypdf==4.3.1

# Vector database (Phase 3)
qdrant-client==1.11.1

# Embeddings (Phase 3)
sentence-transformers==3.1.1

# requirements.txt — add these lines

# Database (Phase 5)
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.10
aiosqlite==0.20.0

# Auth (Phase 6) — replace passlib with bcrypt directly
bcrypt==4.0.1
python-jose[cryptography]==3.3.0